refactoring as requested by codeclimate
